mockeado del breadcrumb de las catgorias en el back y el front
