+++ src/modules/ProductDetails/productDetail.component.js
@@ -4,6 +4,7 @@ import Button from '../../Components/Button/button'
 import Loading from '../../Components/Loaders/loading.component'
 import Error from '../../Components/Error/error.component'
 import { priceFormatter, decimalSplitter, formatText } from '../../utils'
+import Breadcrumb from '../../Components/Breadcrumb/breadcrumb.component'
 
 const productCondition = {
   new: 'Nuevo',
@@ -11,14 +12,15 @@ const productCondition = {
 }
 
 const ProductDetailView = ({
- sold_quantity,
- condition,
- picture,
- title,
- price = {},
- description,
- isLoading,
- isError
+  sold_quantity,
+  condition,
+  picture,
+  title,
+  price = {},
+  description,
+  categories,
+  isLoading,
+  isError
 }) => {
 
   const InfoSailsTitle = () => (
@@ -35,24 +37,28 @@ const ProductDetailView = ({
   if (isError) return <Error />
 
   return (
-    <div className="product-detail-container">
-      <InfoSailsTitle />
-      <img src={picture} alt={title} />
-      <div className="product-detail-info">
+    <>
+      <Breadcrumb crumbs={categories} />
+
+      <div className="product-detail-container">
         <InfoSailsTitle />
+        <img src={picture} alt={title} />
+        <div className="product-detail-info">
+          <InfoSailsTitle />
 
-        <h3 className="product-detail-info_price">
-          {priceFormatter(price)}
-          {price.decimals > 0 && <span className="price-decimals">{decimalSplitter(price.decimals)}</span>}
-        </h3>
+          <h3 className="product-detail-info_price">
+            {priceFormatter(price)}
+            {price.decimals > 0 && <span className="price-decimals">{decimalSplitter(price.decimals)}</span>}
+          </h3>
 
-        <Button title='Comprar' styleType="buy"></Button>
+          <Button title='Comprar' styleType="buy"></Button>
+        </div>
+        <div className="product-detail-description">
+          <h4 className="product-detail-description_title">Descripción del producto</h4>
+          <p>{niceDescription}</p>
+        </div>
       </div>
-      <div className="product-detail-description">
-        <h4 className="product-detail-description_title">Descripción del producto</h4>
-        <p>{niceDescription}</p>
-      </div>
-    </div>
+    </>
   )
 }
