+++ src/modules/ProductDetails/productDetail.styles.scss
@@ -6,15 +6,15 @@
     height: 100%;
     border-radius: 4px;
     box-shadow: 0 1px 2px 0 rgba(0, 0, 0, .25);
-    margin-top: 76px; // todo quitar luego
     display: grid;
     grid-template-columns: 680px auto;
     justify-items: center;
     color: $darkGray;
 
     img {
-      max-width: 80%;
-      max-height: 100%;
+      width: 100%;
+      max-height: 680px;
+      object-fit: contain;
     }
 
     & > .product-detail-info_mobile {
