+++ src/modules/SearchResults/results.container.js
@@ -4,12 +4,13 @@ import { useGetQuery, useGetData } from '../../Hooks'
 
 const SearchResults = () => {
   const querySearch = useGetQuery()
-  const { data, isLoading } = useGetData(`http://localhost:3001/api/items?q=${querySearch}`, 'items', querySearch)
+  const { data, isLoading } = useGetData(`http://localhost:3001/api/items?q=${querySearch}`, querySearch)
 
   return (
     <SearchResultsView
-      data={data}
+      data={data?.items}
       isLoading={isLoading}
+      categories={data?.categories}
     />
   )
 }
